@@ -3,7 +3,7 @@ description: A Flutter package which provides 360 view of the images with rotati
 version: 1.2.0
 homepage: https://github.com/hd-motion/flutter_imageview_360
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
